backend/requirements.txt
# Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
slowapi>=0.1.9

# Config & validation
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
typing-extensions>=4.10.0

# LangChain / LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# RAG
chromadb>=0.5.0
pypdf>=5.0.0

# Playwright
playwright>=1.49.0

# Async / HTTP
nest-asyncio>=1.6.0
httpx>=0.27.0

# Observability
langfuse>=2.0.0

# Tests
pytest>=8.3.0
pytest-asyncio>=0.24.0
